requirements.txt
telethon>=1.30.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
pillow>=10.0.0